fix: resolve all compilation errors, reduce clippy warnings to 18, 2851 tests passing
- Fix SqlxTransactionManager type inference with URL-based DB dispatch
- Fix ldap3 0.11 API migration (SearchScope→Scope, Mod→Mod<String>)
- Fix darling 0.23 attributes syntax in pre_authorize_macro
- Fix sea-orm 2.0 ActiveModel→ActiveModelEx API change
- Fix Rust 2024 edition ref pattern in let bindings
- Fix inventory::iter API call (needs parentheses)
- Fix LogLevel::INFO→Info, Logger::init_with_config signature
- Fix openapi Response builder pattern (StatusCode, header iteration)
- Replace unwrap() with expect() in production code
- Add #[allow] attributes for acceptable warnings (expect_used, missing_docs)
- Fix multipart parsing: let...else, map_or, split_once
- Fix respond.rs: expect on Response builders
- Fix zipkin.rs: expect on Mutex, Debug impl, slice::from_ref
- Fix redis_cache: dead_code on config fields
- Fix schedule: pass by reference, expect usage
- Fix data-rdbc: remove unnecessary u32 casts
- Fix security/oauth2: merge identical match arms
- Fix config: let...else formatting

+++ crates/hiver-data-rdbc/src/connection.rs
@@ -298,8 +298,8 @@ impl ConnectionPool
             DatabaseType::SQLite =>
             {
                 let sqlx_pool = sqlx::sqlite::SqlitePoolOptions::new()
-                    .max_connections(config.max_size as u32)
-                    .min_connections(config.min_idle as u32)
+                    .max_connections(config.max_size)
+                    .min_connections(config.min_idle)
                     .acquire_timeout(config.connection_timeout)
                     .idle_timeout(config.idle_timeout)
                     .max_lifetime(config.max_lifetime)
